+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+# 0.2.1 (2021-12-14)
+* [DEVELOPMENT-283] Improve performance for wallet reordering & active wallets settings
+* [DEVELOPMENT-275] Disable scrolling on receive page
+
 # 0.2.0 (2021-12-02)
 * Initial simplio-wallet commit